Refactor/#71 smalltalk usecase 폴더 구조 변경
