+++ src/application/usecases/smalltalk/OpenQuestionUsecase.ts
@@ -1,5 +1,5 @@
 import { IOpenQuestionRepository } from "@/domain/repositories/smalltalk/IOpenQuestionRepository";
-import { OpenQuestionDto } from "@/application/usecases/smalltalk/smalltalkOpenQuestion/dto/OpenQuestion";
+import { OpenQuestionDto } from "@/application/usecases/smalltalk/dto/OpenQuestion";
 
 export class OpenQuestionsUsecase {
   constructor(private readonly repository: IOpenQuestionRepository) {}
