+++ src/application/usecases/smalltalk/dto/BalancegameAnswer.ts
