+++ src/application/usecases/smalltalk/dto/OpenQuestion.ts
