+++ src/application/usecases/smalltalk/dto/SmalltalkSubjectDto.ts
