+++ src/types/dropdown.d.ts
@@ -1,4 +1,4 @@
-import { SmalltalkSubjectDto } from "@/application/usecases/smalltalk/smalltalkSubject/dto/SmalltalkSubjectDto";
+import { SmalltalkSubjectDto } from "@/application/usecases/smalltalk/dto/SmalltalkSubjectDto";
 
 export interface DropdownProps {
   options: SmalltalkSubjectDto[];
